Refactor inline style transformer to use String instead of u16 slices
- Updated TransformerData to store source and transformed_source as String.
- Replaced u16 slice operations with String methods for better readability and performance.
- Simplified the query_transform_rules function to accept String parameters directly.
- Adjusted all related functions and tests to accommodate the new String-based implementation.
- Removed unnecessary macros and constants related to u16 handling.
- Updated TypeScript definitions to reflect changes in function signatures.

+++ packages/web-platform/inline-style-parser/src/parse_inline_style.rs
@@ -5,14 +5,13 @@ use crate::{
   utils::cmp_str,
 };
 
-const IMPORTANT_STR: [u16; 9] = [
-  'i' as u16, 'm' as u16, 'p' as u16, 'o' as u16, 'r' as u16, 't' as u16, 'a' as u16, 'n' as u16,
-  't' as u16,
+const IMPORTANT_STR: [u8; 9] = [
+  'i' as u8, 'm' as u8, 'p' as u8, 'o' as u8, 'r' as u8, 't' as u8, 'a' as u8, 'n' as u8, 't' as u8,
 ];
 
 pub struct ParserState<'a, 'b, T: Transformer> {
   transformer: &'b mut T,
-  source: &'a [u16],
+  source: &'a [u8],
   status: usize,
   name_start: usize,
   name_end: usize,
@@ -140,7 +139,7 @@ pub trait Transformer {
   );
 }
 
-pub fn parse_inline_style<T: Transformer>(source: &[u16], transformer: &mut T) {
+pub fn parse_inline_style<T: Transformer>(source: &[u8], transformer: &mut T) {
   let mut parser = ParserState {
     transformer,
     source,
